+++ libvideo/v4l2-tuner.h
@@ -34,7 +34,7 @@
  * @param frequency Frequency value to set
  * @return 0 on success, else LIBVIDEO_ERR_IOCTL
  */
-LIBVIDEO_LOCAL int set_tuner_freq_v4l2(struct video_device *device, unsigned int idx, unsigned int frequency) __attribute__((nonnull (1)));
+LIBVIDEO_LOCAL int set_tuner_freq_v4l2(struct video_device *device, unsigned int idx, unsigned long frequency) __attribute__((nonnull (1)));
 /**
  * Get the v4l2 device's tuner's frequency
  * @param device Device to query
@@ -42,7 +42,7 @@ LIBVIDEO_LOCAL int set_tuner_freq_v4l2(struct video_device *device, unsigned int
  * @param frequency Output for frequency
  * @return 0 on success, else LIBVIDEO_ERR_IOCTL
  */
-LIBVIDEO_LOCAL int get_tuner_freq_v4l2(struct video_device *device, unsigned int idx, unsigned int *frequency) __attribute__((nonnull (1, 3)));
+LIBVIDEO_LOCAL int get_tuner_freq_v4l2(struct video_device *device, unsigned int idx, unsigned long *frequency) __attribute__((nonnull (1, 3)));
 /**
  * Get the v4l2 device's tuner's RSSI and AFC info
  * @param Device device to query
